Update to latest amate data
This is a blind update, no checking was performed and older corrected
data may be lost because too much time has passed and that means
too much data to go through. The scripts don't work entirely right
anyway so whatever. Can consider this a reset as it is simply
a straight up translation of amate's data.